Revert "Merge branch 'p/bmf-status-connected' into p/failing-transactions"
This reverts commit d3e543c4c3e6f03cc74087b05c268d4449ba1689, reversing
changes made to 960001ce00bbc1c99afb6ca6f697748bd6d944f6.
