+++ cmd/node.go
@@ -85,9 +85,6 @@ func (node *Node) Start(exitChannels ...*chan struct{}) {
 		node.Params.EnableRegtest()
 	}
 
-	// Set the frequency of mempool's database backup.
-	node.Params.MempoolBackupTimeMilliseconds = node.Config.MempoolBackupTimeMilliseconds
-
 	// Validate params
 	validateParams(node.Params)
 	// This is a bit of a hack, and we should deprecate this. We rely on GlobalDeSoParams static variable in only one
