+++ go.mod
@@ -15,7 +15,6 @@ require (
 	github.com/deso-protocol/go-deadlock v1.0.0
 	github.com/deso-protocol/go-merkle-tree v1.0.0
 	github.com/dgraph-io/badger/v3 v3.2103.5
-	github.com/emirpasic/gods v1.18.1
 	github.com/ethereum/go-ethereum v1.9.25
 	github.com/fatih/color v1.13.0
 	github.com/gernest/mention v2.0.0+incompatible
