+++ lib/block_view_derived_key_test.go
@@ -877,7 +877,6 @@ func TestBalanceModelAuthorizeDerivedKey(t *testing.T) {
 	// We need to set the block height here to 7 so that encoder migrations have the proper version and heights.
 	// Otherwise, the access groups and associations migrations do not run when encoding Utxo Operations.
 	DeSoTestnetParams.ForkHeights.BalanceModelBlockHeight = 7
-	DeSoTestnetParams.ForkHeights.ProofOfStake1StateSetupBlockHeight = 7
 	TestAuthorizeDerivedKeyWithTransactionSpendingLimitsAccessGroups(t)
 }
